fix: Remove redundant call for permissive mode
