+++ src/main/java/com/resend/services/batch/model/CreateBatchEmailsResponse.java
@@ -1,12 +1,17 @@
 package com.resend.services.batch.model;
 
+import com.fasterxml.jackson.annotation.JsonProperty;
+
 import java.util.List;
 
 /**
- * Response for batch emails sent in strict mode.
- * In strict mode, either all emails are sent successfully or the entire batch fails.
+ * Response for batch emails sent in permissive mode.
+ * In permissive mode, partial success is allowed and validation errors are returned.
  */
 public class CreateBatchEmailsResponse extends AbstractBatchEmailsResponse {
+
+    @JsonProperty("errors")
+    private List<BatchError> errors;
 
     /**
      * Default constructor.
@@ -16,12 +21,41 @@ public class CreateBatchEmailsResponse extends AbstractBatchEmailsResponse {
     }
 
     /**
-     * Constructor with data.
+     * Constructor with data and errors.
      *
      * @param data A list of successfully created batch emails.
+     * @param errors A list of validation errors.
      */
-    public CreateBatchEmailsResponse(final List<BatchEmail> data) {
+    public CreateBatchEmailsResponse(final List<BatchEmail> data, final List<BatchError> errors) {
         super(data);
+        this.errors = errors;
+    }
+
+    /**
+     * Get the list of validation errors.
+     *
+     * @return A list of batch email errors, never null (but may be empty).
+     */
+    public List<BatchError> getErrors() {
+        return errors;
+    }
+
+    /**
+     * Check if the response has any validation errors.
+     *
+     * @return true if there are validation errors, false otherwise.
+     */
+    public boolean hasErrors() {
+        return errors != null && !errors.isEmpty();
+    }
+
+    /**
+     * Get the total number of failed emails.
+     *
+     * @return The count of emails that failed validation.
+     */
+    public int getErrorCount() {
+        return errors != null ? errors.size() : 0;
     }
 
 }
